LayoutTests/platform/wk2/storage/indexeddb/mozilla/key-requirements-inline-and-passed-expected.txt

Test IndexedDB's behavior adding inline and passed key simultaneously

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('baz', { keyPath: 'id' });
Expecting exception from objectStore.add({id: 5}, 5);
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: The data provided does not meet requirements.
PASS successfullyParsed is true

TEST COMPLETE
